+++ API/Controllers/ProjectController.cs
@@ -24,14 +24,11 @@ using Microsoft.AspNetCore.Mvc;
 using Models;
 using Models.Defaults;
 using Services.Services;
-using System;
 using System.Collections.Generic;
-using System.Linq;
 using System.Threading.Tasks;
 
 namespace API.Controllers
 {
-
     /// <summary>
     ///     This controller handles the CRUD projects
     /// </summary>
@@ -39,7 +36,6 @@ namespace API.Controllers
     [ApiController]
     public class ProjectController : ControllerBase
     {
-
         private readonly IMapper mapper;
         private readonly IProjectService projectService;
         private readonly IUserService userService;
@@ -64,8 +60,8 @@ namespace API.Controllers
         [HttpGet]
         public async Task<IActionResult> GetAllProjects()
         {
-            List<Project> projects = await projectService.GetAllWithUsersAsync();
-            if(!projects.Any())
+            List<Project> projects = await projectService.GetAllWithUsersAsync().ConfigureAwait(false);
+            if(projects.Count == 0)
             {
                 ProblemDetails problem = new ProblemDetails
                 {
@@ -78,7 +74,6 @@ namespace API.Controllers
 
             return Ok(mapper.Map<IEnumerable<Project>, IEnumerable<ProjectResourceResult>>(projects));
         }
-
 
         /// <summary>
         ///     Get a project.
@@ -98,7 +93,7 @@ namespace API.Controllers
                 return BadRequest(problem);
             }
 
-            Project project = await projectService.FindWithUserAndCollaboratorsAsync(projectId);
+            Project project = await projectService.FindWithUserAndCollaboratorsAsync(projectId).ConfigureAwait(false);
             if(project == null)
             {
                 ProblemDetails problem = new ProblemDetails
@@ -132,9 +127,7 @@ namespace API.Controllers
                 return BadRequest(problem);
             }
             Project project = mapper.Map<ProjectResource, Project>(projectResource);
-
-            User user = await userService.GetUserAsync(projectResource.UserId);
-            project.User = user;
+            project.User = await HttpContext.GetContextUser(userService).ConfigureAwait(false);
             try
             {
                 projectService.Add(project);
@@ -162,7 +155,7 @@ namespace API.Controllers
         [Authorize]
         public async Task<IActionResult> UpdateProject(int projectId, [FromBody] ProjectResource projectResource)
         {
-            Project project = await projectService.FindAsync(projectId);
+            Project project = await projectService.FindAsync(projectId).ConfigureAwait(false);
             if(project == null)
             {
                 ProblemDetails problem = new ProblemDetails
@@ -176,9 +169,8 @@ namespace API.Controllers
 
             mapper.Map(projectResource, project);
 
-            string identity = HttpContext.User.GetStudentId(HttpContext);
-            bool isAllowed = userService.UserHasScope(identity, nameof(Defaults.Scopes.ProjectWrite));
-            User user = await userService.GetUserByIdentityIdAsync(identity);
+            User user = await HttpContext.GetContextUser(userService).ConfigureAwait(false);
+            bool isAllowed = userService.UserHasScope(user.IdentityId, nameof(Defaults.Scopes.ProjectWrite));
 
             if(!(project.UserId == user.Id || isAllowed))
             {
@@ -188,7 +180,7 @@ namespace API.Controllers
                     Detail = "The user is not allowed to edit the project.",
                     Instance = "2E765D18-8EBC-4117-8F9E-B800E8967038"
                 };
-                return BadRequest(problem);
+                return Unauthorized(problem);
             }
 
             projectService.Update(project);
@@ -204,7 +196,7 @@ namespace API.Controllers
         [Authorize]
         public async Task<IActionResult> DeleteProject(int projectId)
         {
-            Project project = await projectService.FindAsync(projectId);
+            Project project = await projectService.FindAsync(projectId).ConfigureAwait(false);
             if(project == null)
             {
                 ProblemDetails problem = new ProblemDetails
@@ -216,9 +208,8 @@ namespace API.Controllers
                 return NotFound(problem);
             }
             
-            string identity = HttpContext.User.GetStudentId(HttpContext);
-            bool isAllowed = userService.UserHasScope(identity, nameof(Defaults.Scopes.ProjectWrite));
-            User user = await userService.GetUserByIdentityIdAsync(identity);
+            User user = await HttpContext.GetContextUser(userService).ConfigureAwait(false);
+            bool isAllowed = userService.UserHasScope(user.IdentityId, nameof(Defaults.Scopes.ProjectWrite));
 
             if(!(project.UserId == user.Id || isAllowed))
             {
@@ -228,10 +219,10 @@ namespace API.Controllers
                     Detail = "The user is not allowed to delete the project.",
                     Instance = "D0363680-5B4F-40A1-B381-0A7544C70164"
                 };
-                return BadRequest(problem);
+                return Unauthorized(problem);
             }
 
-            await projectService.RemoveAsync(projectId);
+            await projectService.RemoveAsync(projectId).ConfigureAwait(false);
             projectService.Save();
             return Ok();
         }