+++ API/Extensions/AuthorizeScopeAttribute.cs
@@ -27,31 +27,30 @@ using System.Security.Claims;
 
 namespace API.Extensions
 {
-
     /// <summary>
     ///     Easily grab scopes
     /// </summary>
     public class AuthorizeScopeAttribute : TypeFilterAttribute
     {
-
         /// <summary>
+        /// Authorization attribute that is depricated.
         /// </summary>
         /// <param name="claimValue"></param>
         public AuthorizeScopeAttribute(string claimValue) : base(typeof(AuthorizeScopeFilter))
         {
             Arguments = new object[] {new Claim("scope", claimValue)};
         }
-
     }
 
     /// <summary>
+    /// Depricated attribute filter.
     /// </summary>
     public class AuthorizeScopeFilter : IAuthorizationFilter
     {
         private readonly Claim claim;
 
-
         /// <summary>
+        /// Filter constructor.
         /// </summary>
         /// <param name="claim"></param>
         public AuthorizeScopeFilter(Claim claim)
@@ -60,12 +59,12 @@ namespace API.Extensions
         }
 
         /// <summary>
+        /// On authorization check, check for the scopes.
         /// </summary>
         /// <param name="context"></param>
         public void OnAuthorization(AuthorizationFilterContext context)
         {
-
-            var userService = context.HttpContext.RequestServices.GetService(typeof(UserService)) as UserService;        
+            UserService userService = context.HttpContext.RequestServices.GetService(typeof(UserService)) as UserService;
             bool hasClaim = context.HttpContext.User.Claims.Any(c => c.Type == claim.Type && c.Value == claim.Value);
 
             //Get all scopes from the user
@@ -109,7 +108,5 @@ namespace API.Extensions
                 context.Result = new ForbidResult();
             }
         }
-
     }
-
 }