+++ IdentityServer/Configuration/IdentityConfig.cs
@@ -50,7 +50,9 @@ namespace IdentityServer.Configuration
                         new Scope(nameof(Defaults.Scopes.UserWrite)),
                         new Scope(nameof(Defaults.Scopes.UserRead)),
                         new Scope(nameof(Defaults.Scopes.HighlightWrite)),
-                        new Scope(nameof(Defaults.Scopes.HighlightRead))
+                        new Scope(nameof(Defaults.Scopes.HighlightRead)),
+                        new Scope(nameof(Defaults.Scopes.EmbedWrite)),
+                        new Scope(nameof(Defaults.Scopes.EmbedRead)),
                     }
                 }
             };
@@ -77,7 +79,9 @@ namespace IdentityServer.Configuration
                                nameof(Defaults.Scopes.UserWrite),
                                nameof(Defaults.Scopes.UserRead),
                                nameof(Defaults.Scopes.HighlightRead),
-                               nameof(Defaults.Scopes.HighlightWrite)
+                               nameof(Defaults.Scopes.HighlightWrite),
+                               nameof(Defaults.Scopes.EmbedWrite),
+                               nameof(Defaults.Scopes.EmbedRead)
                            }
                        },
 
@@ -106,7 +110,6 @@ namespace IdentityServer.Configuration
                                 IdentityServerConstants.StandardScopes.Profile,
                                 IdentityServerConstants.StandardScopes.Email,
                                 "dex-api"
-
                             },
                            AllowAccessTokensViaBrowser = true
 
@@ -124,15 +127,9 @@ namespace IdentityServer.Configuration
                            PostLogoutRedirectUris = config.Swagger.PostLogoutUrisSwagger,
                            AllowedScopes = new List<string>
                             {
-
                                 IdentityServerConstants.StandardScopes.OpenId,
                                 IdentityServerConstants.StandardScopes.Profile,
-                                nameof(Defaults.Scopes.ProjectRead),
-                                nameof(Defaults.Scopes.ProjectWrite),
-                                nameof(Defaults.Scopes.UserWrite),
-                                nameof(Defaults.Scopes.UserRead),
-                                nameof(Defaults.Scopes.HighlightRead),
-                                nameof(Defaults.Scopes.HighlightWrite)
+                                "dex-api",
                             },
                        }
                    };